move fixing of filenames into utils
